Update Jan 19 2019
Update to preserve changes to this repository during fall 2018. Documenting this is important, but saving this before it changes further is even more so! Especially important are eraT/eraTz (which begin development on ERA functions) and mp4maker (which preserves the settings used to make the Hueholt et al. AMS 2019 poster).